refactor(shell): move shell detection logic from handler to infra
handler/shell.rs is now a thin layer (9 lines) that delegates to
infra::shell_detect. AvailableShell struct and all detection logic
(find_pwsh_path, load_*_shells, push_shell) live in infra.
