+++ src-tauri/crates/infra/src/lib.rs
@@ -5,6 +5,7 @@ pub mod git;
 pub mod logger;
 pub mod no_window;
 pub mod pty;
+pub mod shell_detect;
 pub mod shell_init;
 pub mod slug;
 pub mod watcher;
